task 14

+++ 14-binary_tree_balance.c
@@ -12,8 +12,8 @@ int binary_tree_balance(const binary_tree_t *tree)
 	if (!tree)
 		return (0);
 
-	left_h = binary_tree_height(tree->left);
-	right_h = binary_tree_height(tree->right);
+	left_h = (int)binary_tree_height(tree->left);
+	right_h = (int)binary_tree_height(tree->right);
 
 	return (left_h - right_h);
 }
